pyproject.toml
[project]
name = "demolangextract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langextract>=1.0.5",
    "python-dotenv>=1.1.1",
]
